+++ mex/extractors/consent_mailer/extract.py
@@ -28,12 +28,16 @@ def extract_consents_for_persons(
     """Get consents for ldap persons."""
     connector = BackendApiConnector.get()
     person_ids = [str(person.identifier) for person in person_items]
+
+    if not person_ids:
+        return []
+
     return cast(
         "list[MergedConsent]",
         list(
             connector.fetch_all_merged_items(
                 entity_type=["MergedConsent"],
-                reference_field="hasDataSubject",
+                reference_field="hasDataSubject" if person_ids else None,
                 referenced_identifier=person_ids,
             )
         ),