+++ mex/extractors/consent_mailer/settings.py
@@ -1,5 +1,4 @@
-from pydantic import AnyUrl, Field, SecretStr
-from pydantic_core import Url
+from pydantic import Field
 
 from mex.common.settings import BaseSettings
 from mex.common.types import AssetsPath
@@ -19,4 +18,9 @@ class ConsentMailerSettings(BaseSettings):
             "Path to the directory with the jinja template file containing the"
             " email body template, absolute path or relative to `assets_dir`."
         ),
+    )
+    smtp_server: str = Field(
+        "localhost:1025",
+        description="Address and port (<address>:<port>) of the SMTP server to use for"
+        " the consent mailer.",
     )