+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - finished consent-mailer pipeline
+- settings for the content-mailer
+  - smtp server address for sending mails
+  - mailHog api endpoint url for testing mail sending
 
 ### Changes